python-app/requirements.txt
numpy==2.0.2
pandas==2.2.3
python-dateutil==2.9.0.post0
pytz==2024.2
six==1.17.0
tzdata==2024.2
mysqlclient==2.1.1
mysql-connector==2.2.9
mysql-connector-python==9.1.0